src/integrations/slack/requirements.txt
# Slack Integration Dependencies

# Core Python Libraries
asyncio==3.4.3
typing==3.7.4.3

# Slack SDK
slack-sdk==3.26.0
aiohttp==3.9.3

# Environment and Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Utilities
regex==2023.10.3

# Optional: Logging and Monitoring
structlog==24.1.0

# Optional: Database Integration (to be added in future versions)
# sqlalchemy==2.0.25
# asyncpg==0.29.0

# Optional: Security
cryptography==41.0.7

# Development and Debugging
ipdb==0.13.13

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Type Checking
mypy==1.8.0
types-python-dateutil==2.8.19.20240106

# Linting and Code Quality
flake8==7.0.0
black==24.1.1
isort==5.13.2
